rename TestWorkspaceCreator to TestWorkspaceManager
