+++ tests/org.eclipse.n4js.ide.tests/src/org/eclipse/n4js/ide/tests/builder/IncrementalBuilderPackageJsonTest.xtend
@@ -24,7 +24,7 @@ class IncrementalBuilderPackageJsonTest extends AbstractIncrementalBuilderTest {
 
 	@Test
 	def void testChangeInPackageJson() throws IOException {
-		workspaceCreator.createTestProjectOnDisk(
+		testWorkspaceManager.createTestProjectOnDisk(
 			"Main" -> '''
 				console.log("hello");
 			'''
